+++ app/src/main/java/org/simple/clinic/home/HomeScreenEffect.kt
@@ -1,0 +1,3 @@
+package org.simple.clinic.home
+
+sealed class HomeScreenEffect